+++ core/domain/src/main/java/com/wansenai/bo/warehouse/AllotStockDataExportBO.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansenai.bo;
+package com.wansenai.bo.warehouse;
 
 import com.wansenai.utils.excel.ExcelExport;
 import lombok.AllArgsConstructor;
@@ -20,51 +20,48 @@ import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
 
-/**
- * 组装单和拆卸单导出数据BO (AssembleStockDataExportBO) 共存在这一个类中
- */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AssembleStockDataExportBO {
+public class AllotStockDataExportBO {
 
-    @ExcelExport("单据编号")
+    @ExcelExport(value = "调拨单号")
     private String receiptNumber;
 
-    @ExcelExport("商品类型")
-    private String type;
-
-    @ExcelExport("仓库")
+    @ExcelExport(value = "调出方仓库")
     private String warehouseName;
 
-    @ExcelExport("条码")
+    @ExcelExport(value = "调入方仓库")
+    private String otherWarehouseName;
+
+    @ExcelExport(value = "条码")
     private String barCode;
 
-    @ExcelExport("商品名称")
+    @ExcelExport(value = "商品名称")
     private String productName;
 
-    @ExcelExport("规格")
+    @ExcelExport(value = "规格")
     private String productStandard;
 
-    @ExcelExport("型号")
+    @ExcelExport(value = "型号")
     private String productModel;
 
-    @ExcelExport("单位")
+    @ExcelExport(value = "单位")
     private String productUnit;
 
-    @ExcelExport("库存")
+    @ExcelExport(value = "库存")
     private Integer stock;
 
-    @ExcelExport("商品数量")
+    @ExcelExport(value = "数量")
     private Integer productNumber;
 
-    @ExcelExport("单价")
+    @ExcelExport(value = "单价")
     private BigDecimal unitPrice;
 
-    @ExcelExport("金额")
+    @ExcelExport(value = "金额")
     private BigDecimal amount;
 
-    @ExcelExport("备注")
+    @ExcelExport(value = "备注")
     private String remark;
 }
